pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "controltower"
version = "0.1.0"
description = "Control Tower + Obsidian export layer for ScheduleLab and ProfitIntel."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "jinja2>=3.1,<4.0",
  "pydantic>=2.10,<3.0",
  "pyyaml>=6.0,<7.0",
  "uvicorn>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=9.0,<10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

